Remove intake catalog from Read module (#438)
* delete is2cat.py and references

* remove intake and related modules
