+++ doc/source/user_guide/documentation/read.rst
@@ -19,7 +19,6 @@ Attributes
 .. autosummary::
    :toctree: ../../_icepyx/
 
-   Read.is2catalog
    Read.vars
