+++ requirements.txt
@@ -7,8 +7,6 @@ h5netcdf
 h5py
 holoviews
 hvplot
-intake
-intake-xarray
 matplotlib
 numpy
 requests
